+++ src/providers/twitch/api/Helix.cpp
@@ -142,25 +142,6 @@ void Helix::getUserFollowers(
                             std::move(failureCallback));
 }
 
-void Helix::getUserFollow(
-    QString userId, QString targetId,
-    ResultCallback<bool, HelixUsersFollowsRecord> successCallback,
-    HelixFailureCallback failureCallback)
-{
-    this->fetchUsersFollows(
-        std::move(userId), std::move(targetId),
-        [successCallback](const auto &response) {
-            if (response.data.empty())
-            {
-                successCallback(false, HelixUsersFollowsRecord());
-                return;
-            }
-
-            successCallback(true, response.data[0]);
-        },
-        std::move(failureCallback));
-}
-
 void Helix::fetchStreams(
     QStringList userIds, QStringList userLogins,
     ResultCallback<std::vector<HelixStream>> successCallback,
@@ -352,50 +333,6 @@ void Helix::getGameById(QString gameId,
             successCallback(games[0]);
         },
         failureCallback);
-}
-
-void Helix::followUser(QString userId, QString targetId,
-                       std::function<void()> successCallback,
-                       HelixFailureCallback failureCallback)
-{
-    QUrlQuery urlQuery;
-
-    urlQuery.addQueryItem("from_id", userId);
-    urlQuery.addQueryItem("to_id", targetId);
-
-    this->makeRequest("users/follows", urlQuery)
-        .type(NetworkRequestType::Post)
-        .onSuccess([successCallback](auto /*result*/) -> Outcome {
-            successCallback();
-            return Success;
-        })
-        .onError([failureCallback](auto /*result*/) {
-            // TODO: make better xd
-            failureCallback();
-        })
-        .execute();
-}
-
-void Helix::unfollowUser(QString userId, QString targetId,
-                         std::function<void()> successCallback,
-                         HelixFailureCallback failureCallback)
-{
-    QUrlQuery urlQuery;
-
-    urlQuery.addQueryItem("from_id", userId);
-    urlQuery.addQueryItem("to_id", targetId);
-
-    this->makeRequest("users/follows", urlQuery)
-        .type(NetworkRequestType::Delete)
-        .onSuccess([successCallback](auto /*result*/) -> Outcome {
-            successCallback();
-            return Success;
-        })
-        .onError([failureCallback](auto /*result*/) {
-            // TODO: make better xd
-            failureCallback();
-        })
-        .execute();
 }
 
 void Helix::createClip(QString channelId,